move metadata to a separate file
